define authorized handler trait
