+++ backend/server/src/extractor/DbConnection.rs
@@ -1,0 +1,28 @@
+use crate::AppState;
+use axum::extract::FromRequestParts;
+use axum::http::request::Parts;
+use axum::http::StatusCode;
+use axum::{async_trait, Extension, RequestPartsExt};
+use rustter_query::OwnedAsyncConnection;
+
+pub struct DbConnection(pub OwnedAsyncConnection);
+//                                \___ what we want to extract
+
+#[async_trait]
+impl<S> FromRequestParts<S> for DbConnection
+where
+    S: Send + Sync,
+{
+    type Rejection = (StatusCode, &'static str);
+
+    async fn from_request_parts(parts: &mut Parts, _: &S) -> Result<Self, Self::Rejection> {
+        let Extension(state) = parts.extract::<Extension<AppState>>().await.unwrap(); // can safely unwrap as we added state to the layer
+        let connection = state.db_pool.get_owned().await.map_err(|_| {
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "failed to connect to database",
+            )
+        })?;
+        Ok(Self(connection))
+    }
+}
