+++ backend/server/src/handler.rs
@@ -1,7 +1,9 @@
+mod post;
 pub mod user;
 
 use crate::error::ApiResult;
-use crate::extractor::DbConnection;
+use crate::extractor::DbConnection::DbConnection;
+use crate::extractor::UserSession::UserSession;
 use crate::AppState;
 use axum::extract::State;
 use axum::response::IntoResponse;
@@ -19,6 +21,18 @@ pub trait PublicApiRequest {
     ) -> ApiResult<Self::Response>;
 }
 
+#[async_trait]
+pub trait AuthorizedApiRequest {
+    type Response: IntoResponse;
+
+    async fn process_request(
+        self,
+        conn: DbConnection,
+        user_session: UserSession,
+        state: AppState,
+    ) -> ApiResult<Self::Response>;
+}
+
 pub async fn with_public_handler<'a, Req>(
     conn: DbConnection,
     State(state): State<AppState>,
@@ -28,4 +42,16 @@ where
     Req: PublicApiRequest + Deserialize<'a>,
 {
     payload.process_request(conn, state).await
+}
+
+pub async fn with_handler<'a, Req>(
+    conn: DbConnection,
+    user_session: UserSession,
+    State(state): State<AppState>,
+    Json(payload): Json<Req>,
+) -> ApiResult<Req::Response>
+where
+    Req: AuthorizedApiRequest + Deserialize<'a>,
+{
+    payload.process_request(conn, user_session, state).await
 }
